Add PoC for exporter implementation
Problem:
As a maintainer of NGF and NIC
I want an exporter library to enable me to send data to a centralized
telemetry collector,
So that I can aggregate and visualize telemetry data to make better
product decisions based on how users are using the products.

Solution:

Note: this is a PoC of the design.

- Add Exporter type Exporter the data to an OpenTelemetry endpoint.
- The passed exported data must implement Exportable interface.
- Add generation tools to generate the implementation of
the Exportable interface for the data and generate the scheme.
- Define common telemetry data points to be used across projects.
- Add a go code example along with the instructions for an e2e demo.

Fixes https://github.com/nginxinc/nginx-gateway-fabric/issues/1318