Remove field name decapitalization in the scheme
Problem:

For consistency with DataFabric examples, we make field names start
with lowercase letter. As a result, we convert the first character
of field name in a go struct to lowercase. For example,
MyField -> myField. However, the generated scheme names for fields
like HTTPRoutes look a bit ugly: HTTPRoutes -> hTTPRoutes

Solution:

This is not a hard requirement for Avro scheme, we can use names like
HTTPRoutes just fine.

long? HTTPRoutes = null;

CLOSES -- https://github.com/nginxinc/telemetry-exporter/issues/33