Fix missing attribute sent along with the data
Problem:

F5 telemetry service requires the dataType attribute present, which
must be equal to the `@df_datatype` annotation in the scheme.

Solution:

Add dataType attribute to the generated Attributes() method.

Note: the attribute only required for the struct for which we generate
the scheme. In the Attribute() method of any embedded structs,
dataType attribute must not be set.

CLOSES https://github.com/nginxinc/telemetry-exporter/issues/39