Revert Go version to 1.22.0 (#120)
This is a library so it will be compiled with the version specified in
the project that imports it. This also is treated by Go as a minimum
suggested version and we should only increase it if we need new features
from the language. Bumping this to the latest version only means that
the projects importing it will be forced to update to the same version.